+++ Telegram/ViewModels/Delegates/IMessageDelegate.cs
@@ -50,8 +50,8 @@ namespace Telegram.ViewModels.Delegates
         void OpenLocation(Location location, string title);
         void OpenGame(MessageViewModel message);
         void OpenInlineButton(MessageViewModel message, InlineKeyboardButton button);
-        void OpenMedia(MessageViewModel message, FrameworkElement target, int timestamp = 0);
-        void OpenPaidMedia(MessageViewModel message, PaidMedia media, FrameworkElement target, int timestamp = 0);
+        void OpenMedia(MessageViewModel message, FrameworkElement target, double timestamp = 0);
+        void OpenPaidMedia(MessageViewModel message, PaidMedia media, FrameworkElement target, double timestamp = 0);
         void PlayMessage(MessageViewModel message);
         bool RecognizeSpeech(MessageViewModel message);
